+++ src/capturer/simple.rs
@@ -58,6 +58,13 @@ impl Capturer for SimpleCapturer<'_> {
     &self.pointer_shape_buffer[..self.pointer_shape_buffer_size]
   }
 
+  fn pointer_shape_updated(&self) -> bool {
+    self.pointer_shape_buffer_size != self.last_pointer_shape_buffer_size || {
+      let len = self.pointer_shape_buffer_size;
+      self.pointer_shape_buffer[..len] != self.last_pointer_shape_buffer[..len]
+    }
+  }
+
   fn desc(&self) -> Result<DXGI_OUTPUT_DESC> {
     self.ctx.desc()
   }