+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - **_Breaking Change_**: remove `DuplicationContext.acquire_next_frame`, add `DuplicationContext.next_frame/next_frame_with_pointer_shape`.
 - **_Breaking Change_**: remove `DuplicationContext.get_pointer`, add `DuplicationContext.capture_frame_with_pointer_shape`.
 - **_Breaking Change_**: remove `Capturer.get_pointer`, add `Capturer.capture_with_pointer_shape/pointer_shape_buffer`.
+- Feat: add `Capturer.pointer_shape_updated`.
 
 ## v0.3.0